Web app can know see trucks and truck details

+++ src/Frontends/Mvc/Web/Models/View/BasketViewModel.cs
@@ -7,6 +7,6 @@ namespace Web.Models.View
     {
         public Guid BasketId { get; set; }
         public List<BasketLineViewModel> BasketLines { get; set; }
-        public int ShoppingCartTotal { get; set; }
+        public decimal ShoppingCartTotal { get; set; }
     }
 }
